handled more of the compiler warnings. updated the coverage reports.

+++ src/algorithms/misc/MiscMath.java
@@ -602,7 +602,7 @@ public class MiscMath {
             double diffX = x.get(i) - avgX;
             sumX += (diffX * diffX);
         }
-        double stdDevX = (Math.sqrt((double)sumX/((double)length - 1.)));
+        double stdDevX = (Math.sqrt(sumX/((double)length - 1.)));
         
         return new double[]{avgX, stdDevX};
     }
